refactor(migration): parseXML.js を TypeScript 化
src/migration 配下で最後に残っていた JS ファイルを TS へ変換。fast-xml-parser
が返す生の XML 構造に対して RawPackage 等の型を定義し、PackageInfo /
PackagesList / getPackages の公開シグネチャに型を付けた。ロジック自体は
変更しておらず、サンプル XML でのパース結果も変換前後で一致することを確認済み。
